[Data Usage] Remove joi leaks into the browser (#232355)
## Summary

Upgrading `joi` (https://github.com/elastic/kibana/pull/232118) showed
that the Data Usage plugin is leaking it to the browser (increasing the
bundle size).

Nothing wrong about using `@kbn/config-schema` on the browser. However,
I noticed that this plugin only cared about the types. Since types can
be imported from the server, there is no need to have them in the
`common` directory and risk leaking them again.

### Checklist

Check the PR satisfies following conditions.

Reviewers should verify this PR satisfies this list as well.

- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios
- [x] This was checked for breaking HTTP API changes, and any breaking
changes have been approved by the breaking-change committee. The
`release_note:breaking` label should be applied in these situations.
- [x] The PR description includes the appropriate Release Notes section,
and the correct `release_note:*` label is applied per the
[guidelines](https://www.elastic.co/guide/en/kibana/master/contributing.html#kibana-release-notes-process)
- [x] Review the [backport
guidelines](https://docs.google.com/document/d/1VyN5k91e5OVumlc0Gb9RPa3h1ewuPE705nRtioPiTvY/edit?usp=sharing)
and apply applicable `backport:*` labels.

(cherry picked from commit dc6f3a441e5ab74fd7a2cedb91803397854d2012)

# Conflicts:
#	x-pack/platform/plugins/private/data_usage/common/rest_types/data_streams.ts
#	x-pack/platform/plugins/private/data_usage/public/hooks/use_get_data_streams.ts
#	x-pack/platform/plugins/private/data_usage/public/types.ts
#	x-pack/platform/plugins/private/data_usage/server/routes/internal/usage_metrics_handler.ts
#	x-pack/platform/plugins/private/data_usage/server/services/autoops_api.ts
#	x-pack/platform/test/serverless/api_integration/test_suites/data_usage/mock_api.ts
#	x-pack/platform/test/serverless/api_integration/test_suites/data_usage/tests/data_streams.ts
#	x-pack/platform/test/serverless/api_integration/test_suites/data_usage/tests/data_streams_privileges.ts
#	x-pack/platform/test/serverless/api_integration/test_suites/data_usage/tests/metrics.ts
