+++ x-pack/platform/plugins/private/data_usage/common/rest_types/index.ts
@@ -6,4 +6,3 @@
  */
 
 export * from './usage_metrics';
-export * from './data_streams';
