+++ x-pack/platform/plugins/private/data_usage/common/rest_types/usage_metrics.ts
@@ -5,8 +5,6 @@
  * 2.0.
  */
 
-import { schema, type TypeOf } from '@kbn/config-schema';
-
 // note these should be sorted alphabetically as we sort the URL params on the browser side
 // before making the request, else the cache key will be different and that would invoke a new request
 export const DEFAULT_METRIC_TYPES = ['ingest_rate', 'storage_retained'] as const;
@@ -54,89 +52,3 @@ export const METRIC_TYPE_UI_OPTIONS_VALUES_TO_API_MAP = Object.freeze<Record<str
 // type guard for MetricTypes
 export const isMetricType = (type: string): type is MetricTypes =>
   METRIC_TYPE_VALUES.includes(type as MetricTypes);
-
-// @ts-ignore
-const isValidMetricType = (value: string) => METRIC_TYPE_VALUES.includes(value);
-
-const DateSchema = schema.string({
-  minLength: 1,
-  validate: (v) => (v.trim().length ? undefined : 'Date ISO string must not be empty'),
-});
-
-const metricTypesSchema = schema.oneOf(
-  // @ts-expect-error TS2769: No overload matches this call
-  METRIC_TYPE_VALUES.map((metricType) => schema.literal(metricType)) // Create a oneOf schema for the keys
-);
-export const UsageMetricsRequestSchema = schema.object({
-  from: DateSchema,
-  to: DateSchema,
-  metricTypes: schema.arrayOf(schema.string(), {
-    minSize: 1,
-    validate: (values) => {
-      const trimmedValues = values.map((v) => v.trim());
-      if (trimmedValues.some((v) => !v.length)) {
-        return '[metricTypes] list cannot contain empty values';
-      } else if (trimmedValues.some((v) => !isValidMetricType(v))) {
-        return `must be one of ${METRIC_TYPE_VALUES.join(', ')}`;
-      }
-    },
-  }),
-  dataStreams: schema.arrayOf(schema.string(), {
-    validate: (values) => {
-      if (values.map((v) => v.trim()).some((v) => !v.length)) {
-        return 'list cannot contain empty values';
-      }
-    },
-  }),
-});
-
-export type UsageMetricsRequestBody = TypeOf<typeof UsageMetricsRequestSchema>;
-
-export const UsageMetricsResponseSchema = {
-  body: () =>
-    schema.recordOf(
-      metricTypesSchema,
-      schema.arrayOf(
-        schema.object({
-          name: schema.string(),
-          error: schema.nullable(schema.string()),
-          data: schema.arrayOf(
-            schema.object({
-              x: schema.number(),
-              y: schema.number(),
-            })
-          ),
-        })
-      )
-    ),
-};
-
-export type UsageMetricsResponseSchemaBody = Partial<Record<MetricTypes, MetricSeries[]>>;
-
-export type MetricSeries = TypeOf<typeof UsageMetricsResponseSchema.body>[MetricTypes][number];
-
-export const UsageMetricsAutoOpsResponseSchema = {
-  body: () =>
-    schema.recordOf(
-      metricTypesSchema,
-      schema.arrayOf(
-        schema.object({
-          name: schema.string(),
-          error: schema.nullable(schema.string()),
-          data: schema.maybe(
-            schema.nullable(
-              schema.arrayOf(schema.arrayOf(schema.number(), { minSize: 2, maxSize: 2 }))
-            )
-          ),
-        })
-      )
-    ),
-};
-
-export type UsageMetricsAutoOpsResponseMetricSeries = TypeOf<
-  typeof UsageMetricsAutoOpsResponseSchema.body
->[MetricTypes][number];
-
-export type UsageMetricsAutoOpsResponseSchemaBody = Partial<
-  Record<MetricTypes, UsageMetricsAutoOpsResponseMetricSeries[]>
->;
