+++ x-pack/platform/plugins/private/data_usage/public/app/components/data_usage_metrics.tsx
@@ -12,7 +12,8 @@ import { i18n } from '@kbn/i18n';
 import { Charts } from './charts';
 import { useBreadcrumbs } from '../../utils/use_breadcrumbs';
 import { useKibanaContextForPlugin } from '../../utils/use_kibana';
-import { DEFAULT_METRIC_TYPES, type UsageMetricsRequestBody } from '../../../common/rest_types';
+import { DEFAULT_METRIC_TYPES } from '../../../common/rest_types';
+import type { UsageMetricsRequestBody } from '../../../server/routes/internal/usage_metrics';
 import { PLUGIN_NAME } from '../../translations';
 import { useGetDataUsageMetrics } from '../../hooks/use_get_usage_metrics';
 import { useGetDataUsageDataStreams } from '../../hooks/use_get_data_streams';
