+++ x-pack/platform/plugins/private/data_usage/public/hooks/use_get_data_streams.ts
@@ -8,7 +8,7 @@
 import type { UseQueryOptions, UseQueryResult } from '@tanstack/react-query';
 import { useQuery } from '@tanstack/react-query';
 import type { IHttpFetchError } from '@kbn/core-http-browser';
-import { DataStreamsResponseBodySchemaBody } from '../../common/rest_types';
+import type { DataStreamsResponseBodySchemaBody } from '../../server/routes/internal/data_streams';
 import { DATA_USAGE_DATA_STREAMS_API_ROUTE, DEFAULT_SELECTED_OPTIONS } from '../../common';
 import { useKibanaContextForPlugin } from '../utils/use_kibana';
