+++ x-pack/platform/plugins/private/data_usage/server/routes/internal/data_streams_handler.ts
@@ -9,10 +9,7 @@ import { RequestHandler } from '@kbn/core/server';
 import { DataUsageContext, DataUsageRequestHandlerContext } from '../../types';
 import { errorHandler } from '../error_handler';
 import { getMeteringStats } from '../../utils/get_metering_stats';
-import type {
-  DataStreamsRequestQuery,
-  DataStreamsResponseBodySchemaBody,
-} from '../../../common/rest_types/data_streams';
+import type { DataStreamsRequestQuery, DataStreamsResponseBodySchemaBody } from './data_streams';
 import { NoIndicesMeteringError, NoPrivilegeMeteringError } from '../../errors';
 import { CustomHttpRequestError } from '../../utils';
