+++ x-pack/platform/plugins/private/data_usage/server/routes/internal/usage_metrics.ts
@@ -5,7 +5,9 @@
  * 2.0.
  */
 
-import { UsageMetricsRequestSchema, UsageMetricsResponseSchema } from '../../../common/rest_types';
+import { schema, type TypeOf } from '@kbn/config-schema';
+import { metricTypesSchema } from '../../services/autoops_api';
+import { isMetricType, METRIC_TYPE_VALUES, type MetricTypes } from '../../../common/rest_types';
 import { DATA_USAGE_METRICS_API_ROUTE } from '../../../common';
 import { DataUsageContext, DataUsageRouter } from '../../types';
 
@@ -41,3 +43,56 @@ export const registerUsageMetricsRoute = (
       getUsageMetricsHandler(dataUsageContext)
     );
 };
+
+const DateSchema = schema.string({
+  minLength: 1,
+  validate: (v) => (v.trim().length ? undefined : 'Date ISO string must not be empty'),
+});
+
+export const UsageMetricsRequestSchema = schema.object({
+  from: DateSchema,
+  to: DateSchema,
+  metricTypes: schema.arrayOf(schema.string(), {
+    minSize: 1,
+    validate: (values) => {
+      const trimmedValues = values.map((v) => v.trim());
+      if (trimmedValues.some((v) => !v.length)) {
+        return '[metricTypes] list cannot contain empty values';
+      } else if (trimmedValues.some((v) => !isMetricType(v))) {
+        return `must be one of ${METRIC_TYPE_VALUES.join(', ')}`;
+      }
+    },
+  }),
+  dataStreams: schema.arrayOf(schema.string(), {
+    validate: (values) => {
+      if (values.map((v) => v.trim()).some((v) => !v.length)) {
+        return 'list cannot contain empty values';
+      }
+    },
+  }),
+});
+
+export type UsageMetricsRequestBody = TypeOf<typeof UsageMetricsRequestSchema>;
+
+const UsageMetricsResponseSchema = {
+  body: () =>
+    schema.recordOf(
+      metricTypesSchema,
+      schema.arrayOf(
+        schema.object({
+          name: schema.string(),
+          error: schema.nullable(schema.string()),
+          data: schema.arrayOf(
+            schema.object({
+              x: schema.number(),
+              y: schema.number(),
+            })
+          ),
+        })
+      )
+    ),
+};
+
+export type MetricSeries = TypeOf<typeof UsageMetricsResponseSchema.body>[MetricTypes][number];
+
+export type UsageMetricsResponseSchemaBody = Partial<Record<MetricTypes, MetricSeries[]>>;
