+++ x-pack/platform/plugins/private/data_usage/server/routes/internal/usage_metrics_handler.ts
@@ -8,13 +8,10 @@
 import { chunk } from 'lodash/fp';
 import { RequestHandler } from '@kbn/core/server';
 import { momentDateParser } from '../../../common/utils';
-import type {
-  MetricTypes,
-  UsageMetricsAutoOpsResponseSchemaBody,
-  UsageMetricsRequestBody,
-  UsageMetricsResponseSchemaBody,
-} from '../../../common/rest_types';
-import { DataUsageContext, DataUsageRequestHandlerContext } from '../../types';
+import type { MetricTypes } from '../../../common/rest_types';
+import type { UsageMetricsAutoOpsResponseSchemaBody } from '../../services/autoops_api';
+import type { UsageMetricsRequestBody, UsageMetricsResponseSchemaBody } from './usage_metrics';
+import type { DataUsageContext, DataUsageRequestHandlerContext } from '../../types';
 
 import { errorHandler } from '../error_handler';
 import { CustomHttpRequestError } from '../../utils';
