+++ x-pack/platform/test/serverless/api_integration/test_suites/data_usage/tests/data_streams.ts
@@ -6,7 +6,7 @@
  */
 
 import expect from '@kbn/expect';
-import { DataStreamsResponseBodySchemaBody } from '@kbn/data-usage-plugin/common/rest_types';
+import type { DataStreamsResponseBodySchemaBody } from '@kbn/data-usage-plugin/server/routes/internal/data_streams';
 import { DATA_USAGE_DATA_STREAMS_API_ROUTE } from '@kbn/data-usage-plugin/common';
 import { SupertestWithRoleScopeType } from '../../../services';
 import { FtrProviderContext } from '../../../ftr_provider_context';
